fix: cleaner code
