--- crates/oxc/test.ts
@@ -1,3 +1,0 @@
-console.log("Hello, world!");
-console.log(123);
-// console.log(123);
